Custom mob drops
You can now have custom mob drops, which is configurable in the
config.yml file.
You must use the item name (can use custom items) and split them with
"," (commas) and then when the certain mob dies it will drop one of the
things on that list.
Also some comments in the code